StyledPaneScaffold: rename
