+++ app/src/main/java/page/ooooo/geoshare/ui/LinkScreen.kt
@@ -71,7 +71,7 @@ import page.ooooo.geoshare.lib.billing.Feature
 import page.ooooo.geoshare.lib.geo.CoordinateConverter
 import page.ooooo.geoshare.lib.geo.Geometries
 import page.ooooo.geoshare.lib.geo.Srs
-import page.ooooo.geoshare.ui.components.BasicListDetailScaffold
+import page.ooooo.geoshare.ui.components.StyledListDetailPaneScaffold
 import page.ooooo.geoshare.ui.components.ConfirmationDialog
 import page.ooooo.geoshare.ui.components.FeatureBadgeSmall
 import page.ooooo.geoshare.ui.components.FeatureBadged
@@ -83,7 +83,7 @@ import page.ooooo.geoshare.ui.components.ParagraphText
 import page.ooooo.geoshare.ui.components.ScrollablePane
 import page.ooooo.geoshare.ui.components.SegmentedList
 import page.ooooo.geoshare.ui.components.SegmentedListLabel
-import page.ooooo.geoshare.ui.components.TwoPaneScaffoldDefaults
+import page.ooooo.geoshare.ui.components.StyledPaneScaffoldDefaults
 import page.ooooo.geoshare.ui.theme.AppTheme
 import page.ooooo.geoshare.ui.theme.LocalSpacing
 
@@ -230,7 +230,7 @@ private fun LinkScreen(
     ) {
         // Use BasicListDetailScaffold instead of NavigableBasicListDetailScaffold, because the latter navigates using
         // the navigator when back button is pressed, but we want to do all navigation ourselves using the view model.
-        BasicListDetailScaffold(
+        StyledListDetailPaneScaffold(
             directive = navigator.scaffoldDirective,
             scaffoldState = navigator.scaffoldState,
             listPane = {
@@ -284,7 +284,7 @@ private fun LinkScreen(
                     )
                 }
             },
-            colors = TwoPaneScaffoldDefaults.colors(
+            colors = StyledPaneScaffoldDefaults.colors(
                 wideMainContainerColor = MaterialTheme.colorScheme.surfaceContainer,
             ),
         )
